# openapi.yaml
openapi: 3.0.3
info:
  title: Vetios AI Operations Platform API
  version: '2026-04-05'
  description: Core contract for inference, evaluation, governance, telemetry, webhooks, datasets, and simulations.
servers:
  - url: https://app.vetios.ai
security:
  - bearerAuth: []
    tenantScope: []
paths:
...
  /api/simulate:
    post:
      summary: Run the legacy simulation endpoint with tenant-aware throttling
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SimulateRunRequest' }
      responses:
        '200': { description: Simulation completed, content: { application/json: { schema: { $ref: '#/components/schemas/SimulateRunResponse' } } } }
        '429': { description: Tenant rate limit exceeded, content: { application/json: { schema: { $ref: '#/components/schemas/RateLimitExceeded' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    tenantScope: { type: apiKey, in: header, name: X-Tenant-Scope }
  schemas:
...
    RateLimitExceeded:
      type: object
      properties:
        error: { type: string, enum: [rate_limit_exceeded] }
        tenant_id: { type: string }
        limit: { type: integer }
        window_seconds: { type: integer }
        retry_after_seconds: { type: integer }
...
    SimulateRunRequest:
      type: object
      additionalProperties: true
    SimulateRunResponse:
      type: object
      properties:
        simulation_event_id: { type: string, format: uuid }
        clinical_case_id: { type: string }
        inference_output: { type: object, additionalProperties: true }
        confidence_score: { type: number, nullable: true }
        inference_latency_ms: { type: number }
        simulation: { type: object, additionalProperties: true }
...